Tweak fields in form to use new address history Field type
