+++ controllers/apply/lib/field-types/index.js
@@ -10,6 +10,10 @@ module.exports = {
     CheckboxField: require('./checkbox'),
     SelectField: require('./select'),
     DateField: require('./date'),
+    DayMonthField: require('./day-month'),
+    MonthYearField: require('./month-year'),
     AddressField: require('./address'),
+    AddressHistoryField: require('./address-history'),
     NameField: require('./name'),
+    BudgetField: require('./budget'),
 };
